deleted RaceNetWorkStorage, added 'suspend' to functions. added operator to GetRaceListUseCase.kt
